requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
ChatterBot==1.2.0
chatterbot-corpus==1.2.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.3.2
MarkupSafe==3.0.2
mathparse==0.1.2
nltk==3.9.1
numpy==1.26.1
openai==1.61.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyngrok==7.2.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
threadpoolctl==3.2.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
